fix bug with string sorting and use statistics everywhere we sort
