+++ src/optimizer/statistics/operator/propagate_window.cpp
@@ -16,7 +16,7 @@ unique_ptr<NodeStatistics> StatisticsPropagator::PropagateStatistics(LogicalWind
 			over_expr->partitions_stats.push_back(PropagateExpression(expr));
 		}
 		for (auto &bound_order : over_expr->orders) {
-			bound_order.stats = PropagateExpression(bound_order.expression);
+			PropagateBoundOrder(bound_order);
 		}
 	}
 	return move(node_stats);
